organised backtest and utils file
